+++ app/src/main/kotlin/com/simplemobiletools/calculator/operation/base/UnaryOperation.kt
@@ -1,3 +1,5 @@
 package com.simplemobiletools.calculator.operation.base
 
-open class UnaryOperation protected constructor(protected var value: Double)
+import java.math.BigDecimal
+
+open class UnaryOperation protected constructor(protected var value: BigDecimal)
